+++ tests/conftest.py
@@ -4,6 +4,7 @@ import typing
 import uuid
 
 from pytest import fixture
+from six import PY2, PY3
 
 from nirum.serialize import serialize_record_type, serialize_unboxed_type
 from nirum.deserialize import deserialize_record_type, deserialize_unboxed_type
@@ -11,243 +12,26 @@ from nirum.validate import (validate_unboxed_type, validate_record_type,
                             validate_union_type)
 from nirum.constructs import NameDict, name_dict_type
 
-
-class Token:
-
-    __nirum_inner_type__ = uuid.UUID
-
-    def __init__(self, value: uuid.UUID) -> None:
-        validate_unboxed_type(value, uuid.UUID)
-        self.value = value
-
-    def __eq__(self, other) -> bool:
-        return (isinstance(other, Token) and self.value == other.value)
-
-    def __hash__(self) -> int:
-        return hash(self.value)
-
-    def __nirum_serialize__(self) -> typing.Mapping[str, typing.Any]:
-        return serialize_unboxed_type(self)
-
-    @classmethod
-    def __nirum_deserialize__(
-        cls: type, value: typing.Mapping[str, typing.Any]
-    ) -> 'Token':
-        return deserialize_unboxed_type(cls, value)
-
-    def __hash__(self) -> int:  # noqa
-        return hash((self.__class__, self.value))
+if PY2:
+    nirum_fixture_name = 'tests.py2_nirum'
+elif PY3:
+    nirum_fixture_name = 'tests.py3_nirum'
+else:
+    raise ImportError()
 
 
-class Offset:
-
-    __nirum_inner_type__ = float
-
-    def __init__(self, value):
-        self.value = value
-
-    def __eq__(self, other):
-        return (isinstance(other, Offset) and self.value == other.value)
-
-    def __hash__(self):
-        return hash(self.value)
-
-    def __nirum_serialize__(self):
-        return serialize_boxed_type(self)
-
-    @classmethod
-    def __nirum_deserialize__(cls, value):
-        return deserialize_boxed_type(cls, value)
-
-    def __hash__(self): # noqa
-        return hash(self.value)
-
-
-class Point:
-
-    __slots__ = (
-        'left',
-        'top'
-    )
-    __nirum_record_behind_name__ = 'point'
-    __nirum_field_types__ = {
-        'left': Offset,
-        'top': Offset
-    }
-    __nirum_field_names__ = NameDict([
-        ('left', 'x')
-    ])
-
-    def __init__(self, left, top):
-        self.left = left
-        self.top = top
-        validate_record_type(self)
-
-    def __repr__(self):
-        return '{0.__module__}.{0.__qualname__}({1})'.format(
-            type(self),
-            ', '.join('{}={}'.format(attr, getattr(self, attr))
-                      for attr in self.__slots__)
-        )
-
-    def __eq__(self, other):
-        return isinstance(other, Point) and all(
-            getattr(self, attr) == getattr(other, attr)
-            for attr in self.__slots__
-        )
-
-    def __nirum_serialize__(self):
-        return serialize_record_type(self)
-
-    @classmethod
-    def __nirum_deserialize__(cls, values):
-        return deserialize_record_type(cls, values)
-
-    def __hash__(self):
-        return hash((self.__class__, self.left, self.top))
-
-
-class Shape(object):
-
-    __nirum_union_behind_name__ = 'shape'
-    __nirum_field_names__ = NameDict([
-    ])
-
-    class Tag(enum.Enum):
-        rectangle = 'rectangle'
-        circle = 'circle'
-
-    def __init__(self, *args, **kwargs):
-        raise NotImplementedError(
-            "{0.__module__}.{0.__qualname__} cannot be instantiated "
-            "since it is an abstract class.  Instantiate a concrete subtype "
-            "of it instead.".format(
-                type(self)
-            )
-        )
-
-    def __nirum_serialize__(self):
-        pass
-
-    @classmethod
-    def __nirum_deserialize__(cls, value):
-        pass
-
-
-class Rectangle(Shape):
-
-    __slots__ = (
-        'upper_left',
-        'lower_right'
-    )
-    __nirum_tag__ = Shape.Tag.rectangle
-    __nirum_tag_types__ = {
-        'upper_left': Point,
-        'lower_right': Point
-    }
-    __nirum_tag_names__ = NameDict([])
-
-    def __init__(self, upper_left, lower_right):
-        self.upper_left = upper_left
-        self.lower_right = lower_right
-        validate_union_type(self)
-
-    def __repr__(self):
-        return '{0.__module__}.{0.__qualname__}({1})'.format(
-            type(self),
-            ', '.join('{}={}'.format(attr, getattr(self, attr))
-                      for attr in self.__slots__)
-        )
-
-    def __eq__(self, other):
-        return isinstance(other, Rectangle) and all(
-            getattr(self, attr) == getattr(other, attr)
-            for attr in self.__slots__
-        )
-
-
-class Circle(Shape):
-
-    __slots__ = (
-        'origin',
-        'radius'
-    )
-    __nirum_tag__ = Shape.Tag.circle
-    __nirum_tag_types__ = {
-        'origin': Point,
-        'radius': Offset
-    }
-    __nirum_tag_names__ = NameDict([])
-
-    def __init__(self, origin, radius):
-        self.origin = origin
-        self.radius = radius
-        validate_union_type(self)
-
-    def __repr__(self):
-        return '{0.__module__}.{0.__qualname__}({1})'.format(
-            type(self),
-            ', '.join('{}={}'.format(attr, getattr(self, attr))
-                      for attr in self.__slots__)
-        )
-
-    def __eq__(self, other):
-        return isinstance(other, Circle) and all(
-            getattr(self, attr) == getattr(other, attr)
-            for attr in self.__slots__
-        )
-
-
-class Location:
-    # TODO: docstring
-
-    __slots__ = (
-        'name',
-        'lat',
-        'lng',
-    )
-    __nirum_record_behind_name__ = 'location'
-    __nirum_field_types__ = {
-        'name': typing.Optional[str],
-        'lat': decimal.Decimal,
-        'lng': decimal.Decimal
-    }
-    __nirum_field_names__ = name_dict_type([
-        ('name', 'name'),
-        ('lat', 'lat'),
-        ('lng', 'lng')
-    ])
-
-    def __init__(self, name, lat, lng):
-        self.name = name
-        self.lat = lat
-        self.lng = lng
-        validate_record_type(self)
-
-    def __repr__(self):
-        return '{0.__module__}.{0.__qualname__}({1})'.format(
-            type(self),
-            ', '.join('{}={}'.format(attr, getattr(self, attr))
-                      for attr in self.__slots__)
-        )
-
-    def __eq__(self, other):
-        return isinstance(other, Location) and all(
-            getattr(self, attr) == getattr(other, attr)
-            for attr in self.__slots__
-        )
-
-    def __nirum_serialize__(self):
-        return serialize_record_type(self)
-
-    @classmethod
-    def __nirum_deserialize__(cls, value):
-        return deserialize_record_type(cls, value)
+nirum_fixture = __import__(
+    nirum_fixture_name,
+    globals(),
+    locals(),
+    ['A', 'B', 'C', 'Offset', 'Point', 'Shape', 'Rectangle', 'Circle',
+     'Location']
+)
 
 
 @fixture
-def fx_unboxed_type():
-    return Offset
+def fx_boxed_type():
+    return nirum_fixture.Offset
 
 
 @fixture
@@ -257,7 +41,7 @@ def fx_offset(fx_unboxed_type):
 
 @fixture
 def fx_record_type():
-    return Point
+    return nirum_fixture.Point
 
 
 @fixture
@@ -267,12 +51,12 @@ def fx_point(fx_record_type, fx_unboxed_type):
 
 @fixture
 def fx_circle_type():
-    return Circle
+    return nirum_fixture.Circle
 
 
 @fixture
 def fx_rectangle_type():
-    return Rectangle
+    return nirum_fixture.Rectangle
 
 
 @fixture
@@ -280,105 +64,16 @@ def fx_rectangle(fx_rectangle_type, fx_point):
     return fx_rectangle_type(fx_point, fx_point)
 
 
-class A:
-
-    __nirum_inner_type__ = str
-
-    def __init__(self, value):
-        validate_boxed_type(value, str)
-        self.value = value  # type: Text
-
-    def __eq__(self, other):
-        return (isinstance(other, A) and
-                self.value == other.value)
-
-    def __hash__(self):
-        return hash(self.value)
-
-    def __nirum_serialize__(self):
-        return serialize_boxed_type(self)
-
-    @classmethod
-    def __nirum_deserialize__(cls, value):
-        return deserialize_boxed_type(cls, value)
-
-    def __repr__(self):
-        return '{0.__module__}.{0.__qualname__}({1!r})'.format(
-            type(self), self.value
-        )
-
-
-class B:
-
-    __nirum_inner_type__ = A
-
-    def __init__(self, value):
-        validate_boxed_type(value, A)
-        self.value = value  # type: A
-
-    def __eq__(self, other):
-        return (isinstance(other, B) and
-                self.value == other.value)
-
-    def __hash__(self):
-        return hash(self.value)
-
-    def __nirum_serialize__(self):
-        return serialize_boxed_type(self)
-
-    @classmethod
-    def __nirum_deserialize__(cls, value):
-        return deserialize_boxed_type(cls, value)
-
-    def __repr__(self):
-        return '{0.__module__}.{0.__qualname__}({1!r})'.format(
-            type(self), self.value
-        )
-
-
-class C:
-
-    __nirum_inner_type__ = B
-
-    def __init__(self, value):
-        validate_boxed_type(value, B)
-        self.value = value  # type: B
-
-    def __eq__(self, other):
-        return (isinstance(other, C) and
-                self.value == other.value)
-
-    def __hash__(self):
-        return hash(self.value)
-
-    def __nirum_serialize__(self):
-        return serialize_boxed_type(self)
-
-    @classmethod
-    def __nirum_deserialize__(cls, value):
-        return deserialize_boxed_type(cls, value)
-
-    def __repr__(self):
-        return '{0.__module__}.{0.__qualname__}({1!r})'.format(
-            type(self), self.value
-        )
-
-
 @fixture
-def fx_layered_unboxed_types():
-    return A, B, C
+def fx_layered_boxed_types():
+    return nirum_fixture.A, nirum_fixture.B, nirum_fixture.C
 
 
 @fixture
 def fx_location_record():
-    return Location
+    return nirum_fixture.Location
 
 
 @fixture
 def fx_shape_type():
-    return Shape
-
-
-@fixture
-def fx_token_type():
-    return Token
+    return nirum_fixture.Shape
