+++ tests/validate_test.py
@@ -1,6 +1,7 @@
 import decimal
 
 from pytest import raises
+from six import text_type
 
 from nirum.validate import (validate_unboxed_type, validate_record_type,
                             validate_union_type)
@@ -9,7 +10,7 @@ from nirum.validate import (validate_unboxed_type, validate_record_type,
 def test_validate_unboxed_type():
     assert validate_unboxed_type(3.14, float)
     with raises(TypeError):
-        validate_unboxed_type('hello', float)
+        validate_boxed_type(u'hello', float)
 
 
 def test_validate_record_type(fx_point, fx_record_type, fx_offset,
@@ -37,16 +38,16 @@ def test_validate_union_type(fx_rectangle, fx_rectangle_type, fx_point):
         validate_union_type(fx_rectangle_type(1, 1))
 
 
-def test_validate_layered_unboxed_types(fx_layered_unboxed_types):
-    A, B, C = fx_layered_unboxed_types
-    assert validate_unboxed_type('test', str)
-    assert validate_unboxed_type(A('test'), A)
-    assert validate_unboxed_type(B(A('test')), B)
+def test_validate_layered_boxed_types(fx_layered_boxed_types):
+    A, B, C = fx_layered_boxed_types
+    assert validate_boxed_type(u'test', text_type)
+    assert validate_boxed_type(A(u'test'), A)
+    assert validate_boxed_type(B(A(u'test')), B)
     with raises(TypeError):
-        assert validate_unboxed_type('test', A)
+        assert validate_boxed_type(u'test', A)
 
     with raises(TypeError):
-        assert validate_unboxed_type('test', B)
+        assert validate_boxed_type(u'test', B)
 
     with raises(TypeError):
-        assert validate_unboxed_type(A('test'), B)
+        assert validate_boxed_type(A(u'test'), B)
